+++ src/PSRule.Benchmark/PSRule.cs
@@ -8,21 +8,26 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Management.Automation;
+using System.Net.Http.Headers;
 using System.Reflection;
+using YamlDotNet.Core.Tokens;
 
 namespace PSRule.Benchmark
 {
     public sealed class TargetObject
     {
-        public TargetObject(string name, string message)
+        public TargetObject(string name, string message, string value)
         {
             Name = name;
             Message = message;
+            Value = value;
         }
 
         public string Name { get; private set; }
 
         public string Message { get; private set; }
+
+        public string Value { get; private set; }
     }
 
     /// <summary>
@@ -38,6 +43,9 @@ namespace PSRule.Benchmark
         private InvokeRulePipeline _InvokeIfPipeline;
         private InvokeRulePipeline _InvokeTypePipeline;
         private InvokeRulePipeline _InvokeSummaryPipeline;
+        private InvokeRulePipeline _InvokeWithinPipeline;
+        private InvokeRulePipeline _InvokeWithinBulkPipeline;
+        private InvokeRulePipeline _InvokeWithinLikePipeline;
 
         [GlobalSetup]
         public void Prepare()
@@ -47,6 +55,9 @@ namespace PSRule.Benchmark
             PrepareInvokeIfPipeline();
             PrepareInvokeTypePipeline();
             PrepareInvokeSummaryPipeline();
+            PrepareInvokeWithinPipeline();
+            PrepareInvokeWithinBulkPipeline();
+            PrepareInvokeWithinLikePipeline();
             PrepareTargetObjects();
         }
 
@@ -96,9 +107,41 @@ namespace PSRule.Benchmark
             _InvokeSummaryPipeline = builder.Build();
         }
 
+        private void PrepareInvokeWithinPipeline()
+        {
+            var option = new PSRuleOption();
+            option.Baseline.RuleName = new string[] { "BenchmarkWithin" };
+            var builder = PipelineBuilder.Invoke().Configure(option);
+            builder.Source(new RuleSource[] { new RuleSource(path: GetWithinSource(), moduleName: null) });
+            _InvokeWithinPipeline = builder.Build();
+        }
+
+        private void PrepareInvokeWithinBulkPipeline()
+        {
+            var option = new PSRuleOption();
+            option.Baseline.RuleName = new string[] { "BenchmarkWithinBulk" };
+            var builder = PipelineBuilder.Invoke().Configure(option);
+            builder.Source(new RuleSource[] { new RuleSource(path: GetWithinSource(), moduleName: null) });
+            _InvokeWithinBulkPipeline = builder.Build();
+        }
+
+        private void PrepareInvokeWithinLikePipeline()
+        {
+            var option = new PSRuleOption();
+            option.Baseline.RuleName = new string[] { "BenchmarkWithinLike" };
+            var builder = PipelineBuilder.Invoke().Configure(option);
+            builder.Source(new RuleSource[] { new RuleSource(path: GetWithinSource(), moduleName: null) });
+            _InvokeWithinLikePipeline = builder.Build();
+        }
+
         private string GetSource()
         {
             return Path.Combine(Path.GetDirectoryName(Assembly.GetEntryAssembly().Location), "Benchmark.Rule.ps1");
+        }
+
+        private string GetWithinSource()
+        {
+            return Path.Combine(Path.GetDirectoryName(Assembly.GetEntryAssembly().Location), "Benchmark.Within.Rule.ps1");
         }
 
         private void PrepareTargetObjects()
@@ -109,7 +152,8 @@ namespace PSRule.Benchmark
             while (targetObjects.Count < 1000)
             {
                 r.NextBytes(randomBuffer);
-                var o = new TargetObject(name: targetObjects.Count.ToString(), message: Convert.ToBase64String(randomBuffer));
+                var value = (targetObjects.Count & 1) == 1 ? "Microsoft.Compute/virtualMachines" : "Microsoft.Sql/servers/databases";
+                var o = new TargetObject(name: targetObjects.Count.ToString(), message: Convert.ToBase64String(randomBuffer), value: value);
                 targetObjects.Add(PSObject.AsPSObject(o));
             }
 
@@ -130,6 +174,15 @@ namespace PSRule.Benchmark
 
         [Benchmark]
         public void Get() => _GetPipeline.Process().Consume(new Consumer());
+
+        [Benchmark]
+        public void Within() => _InvokeWithinPipeline.Process(_TargetObject);
+
+        [Benchmark]
+        public void WithinBulk() => _InvokeWithinBulkPipeline.Process(_TargetObject);
+
+        [Benchmark]
+        public void WithinLike() => _InvokeWithinLikePipeline.Process(_TargetObject);
 
         [Benchmark]
         public void DefaultTargetNameBinding()